+++ config/dev.exs
@@ -6,5 +6,4 @@ config :eventstore, EventStore.Storage,
   password: "postgres",
   database: "commanded_dev",
   hostname: "localhost",
-  pool_size: 10,
-  extensions: [{Postgrex.Extensions.Calendar, []}]
+  pool_size: 10