+++ lib/commanded/eventstore/event_store.ex
@@ -21,6 +21,8 @@ defmodule Commanded.EventStore do
       @event_store unquote adapter
     end
   end
+
+  @type start_from :: :origin | :current | integer
   
   @callback append_to_stream(String.t, non_neg_integer, list(EventData.t)) :: :ok | {:error, reason :: term}
 
@@ -30,16 +32,18 @@ defmodule Commanded.EventStore do
   @callback read_stream_forward(String.t, non_neg_integer) :: {:ok, list(RecordedEvent.t)} | {:error, reason :: term}
   @callback read_stream_forward(String.t, non_neg_integer, non_neg_integer) :: {:ok, list(RecordedEvent.t)} | {:error, reason :: term}
   
-  @callback subscribe_to_all_streams(String.t, pid) :: {:ok, subscription :: any}
+  @callback subscribe_to_all_streams(String.t, pid, start_from) :: {:ok, subscription :: any}
+    | {:error, :subscription_already_exists}
+    | {:error, reason :: term}
 
   @callback unsubscribe_from_all_streams(String.t) :: :ok
 
   @callback record_snapshot(SnapshotData.t) :: :ok | {:error, reason :: term}
 
-  @callback ack_events(subscription :: any, non_neg_integer) :: :ok
-
   @callback read_all_streams_forward() :: {:ok, list(RecordedEvent.t)} | {:error, reason :: term}
   @callback read_all_streams_forward(non_neg_integer) :: {:ok, list(RecordedEvent.t)} | {:error, reason :: term}
   @callback read_all_streams_forward(non_neg_integer, non_neg_integer) :: {:ok, list(RecordedEvent.t)} | {:error, reason :: term}
+
+  @callback delete_snapshot(String.t) :: :ok | {:error, reason :: term}
   
 end