+++ test/event/handle_event_test.exs
@@ -1,16 +1,18 @@
 defmodule Commanded.Event.HandleEventTest do
-	use Commanded.StorageCase
+  use Commanded.StorageCase
+  use Commanded.EventStore
 
   import Commanded.Enumerable, only: [pluck: 2]
+  import Commanded.Assertions.EventAssertions
 
   alias Commanded.Event.AppendingEventHandler
   alias Commanded.Helpers.EventFactory
   alias Commanded.ExampleDomain.AccountBalanceHandler
-	alias Commanded.ExampleDomain.BankAccount.Events.{BankAccountOpened,MoneyDeposited}
-
-	test "should be notified of events" do
+  alias Commanded.ExampleDomain.BankAccount.Events.{BankAccountOpened,MoneyDeposited}
+  
+  test "should be notified of events" do
     {:ok, _} = AccountBalanceHandler.start_link
-		{:ok, handler} = Commanded.Event.Handler.start_link("account_balance", AccountBalanceHandler)
+    {:ok, handler} = Commanded.Event.Handler.start_link("account_balance", AccountBalanceHandler)
 
     events = [
       %BankAccountOpened{account_number: "ACC123", initial_balance: 1_000},
@@ -25,7 +27,7 @@ defmodule Commanded.Event.HandleEventTest do
     refute_receive({:ack, _event_id})
 
     assert AccountBalanceHandler.current_balance == 1_050
-	end
+  end
 
   defmodule UninterestingEvent, do: defstruct [field: nil]
 
@@ -55,6 +57,47 @@ defmodule Commanded.Event.HandleEventTest do
 
     assert AccountBalanceHandler.current_balance == 1_050
   end
+
+  test "should ignore events created before the event handler's subscription when starting from `current`" do
+    {:ok, _pid} = AppendingEventHandler.start_link
+
+    stream_uuid = UUID.uuid4
+    initial_events = [%BankAccountOpened{account_number: "ACC123", initial_balance: 1_000}]
+    new_events = [%MoneyDeposited{amount: 50, balance: 1_050}]
+
+    :ok = @event_store.append_to_stream(stream_uuid, 0, Commanded.Event.Mapper.map_to_event_data(initial_events, UUID.uuid4))
+
+    :timer.sleep(400)
+    {:ok, _handler} = Commanded.Event.Handler.start_link("test_event_handler", AppendingEventHandler, start_from: :current)
+
+    :ok = @event_store.append_to_stream(stream_uuid, 1, Commanded.Event.Mapper.map_to_event_data(new_events, UUID.uuid4))
+
+    wait_for_event MoneyDeposited
+    :timer.sleep(200)
+
+    assert AppendingEventHandler.received_events == new_events
+    assert pluck(AppendingEventHandler.received_metadata, :event_id) == [2]
+	end
+
+  test "should receive events created before the event handler's subscription when starting from `origin`" do
+    {:ok, _} = AppendingEventHandler.start_link
+
+    stream_uuid = UUID.uuid4
+    initial_events = [%BankAccountOpened{account_number: "ACC123", initial_balance: 1_000}]
+    new_events = [%MoneyDeposited{amount: 50, balance: 1_050}]
+
+    :ok = @event_store.append_to_stream(stream_uuid, 0, Commanded.Event.Mapper.map_to_event_data(initial_events, UUID.uuid4))
+
+    {:ok, _handler} = Commanded.Event.Handler.start_link("test_event_handler", AppendingEventHandler, start_from: :origin)
+
+    :ok = @event_store.append_to_stream(stream_uuid, 1, Commanded.Event.Mapper.map_to_event_data(new_events, UUID.uuid4))
+
+    wait_for_event MoneyDeposited
+    :timer.sleep(200)
+
+    assert AppendingEventHandler.received_events == initial_events ++ new_events
+    assert pluck(AppendingEventHandler.received_metadata, :event_id) == [1, 2]
+	end
 
 	test "should ignore already seen events" do
     {:ok, _} = AppendingEventHandler.start_link