+++ test/process_managers/resume_process_manager_test.exs
@@ -3,7 +3,7 @@ defmodule Commanded.ProcessManager.ResumeProcessManagerTest do
 
   alias Commanded.Helpers
   alias Commanded.Helpers.Wait
-  alias Commanded.ProcessManagers.ProcessRouter
+  alias Commanded.ProcessManagers.{ProcessRouter,ProcessManagerInstance}
 
   import Commanded.Assertions.EventAssertions
 
@@ -105,9 +105,12 @@ defmodule Commanded.ProcessManager.ResumeProcessManagerTest do
       assert event.status == "start"
     end
 
+    :timer.sleep 100
+
     # wait for process instance to receive event
     Wait.until(fn ->
-      %{status_history: ["start"]} = ProcessRouter.process_state(process_router, process_uuid)
+      process_instance = ProcessRouter.process_instance(process_router, process_uuid)
+      %{status_history: ["start"]} = ProcessManagerInstance.process_state(process_instance)
     end)
 
     Helpers.Process.shutdown(process_router)
@@ -121,9 +124,11 @@ defmodule Commanded.ProcessManager.ResumeProcessManagerTest do
 
     wait_for_event(ProcessResumed, fn event -> event.process_uuid == process_uuid end)
 
-    case ProcessRouter.process_state(process_router, process_uuid) do
+    case ProcessRouter.process_instance(process_router, process_uuid) do
       {:error, :process_manager_not_found} -> flunk("process state not available")
-      state -> assert state.status_history == ["start", "resume"]
+      process_instance ->
+        state = ProcessManagerInstance.process_state(process_instance)
+        assert state.status_history == ["start", "resume"]
     end
   end
 end